Addresses issues

+++ eulerapi.go
@@ -33,7 +33,7 @@ func (s *Server) intSolve(ctx context.Context, number int32) *pb.SolveResponse {
 		}
 	case 3:
 		return &pb.SolveResponse{
-			Answer: int64(s.solve3(ctx, 600851475143)),
+			Answer: s.solve3(ctx, 600851475143),
 		}
 	}
 	return nil
